deepfake_hybrid/requirements.txt
torch
torchvision
timm
numpy
pandas
pyyaml
tqdm
scikit-learn
opencv-python
Pillow
